+++ src/gui/widgets/dialogs/glaxnimate_window.cpp
@@ -600,3 +600,8 @@ void GlaxnimateWindow::update_selection(const std::vector<model::VisualNode*>& s
 {
     return d->selection_changed(selected, deselected);
 }
+
+item_models::DocumentNodeModel* GlaxnimateWindow::node_model() const
+{
+    return &d->document_node_model;
+}